fixed the issue with the active layer tickness
Modified the deviceparams.gridx function to give a warning when the active layer thickness i lower than the thickness of the intermediate layers set in the excel file.
uses a 0.5nm spacing for the layer grid if the thickness is lower than that of the intermediate layers set in the excel file.
introduced a legend function in the dfplot,
added a script to test the effect of the thickeness on the device charachteristics.